% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state(state.num, year)}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{\itemize{
\item numeric identifier for state
}}

\item{year}{\itemize{
\item a calendar year
}}
}
\value{
plot of state which maps accidents for the chosen year
}
\description{
input takes a state number and  a calendar year.
It then:
generates a filename based on the year
reads the data for that year from the file
validates and converts the state.num to an integer\preformatted{IF the state.num is valid it will select the state data for that year
and make a data frame.
IF the dataframe is empty
   it will stop, print a message and return NULL
Otherwise:
   It will plot the Longitude and Latitude of accidents in the state that year
}

Usage

fars_map_state(state.num,year)
}
\examples{
\dontrun{
fars_map_state(1,2013)
}


}
