% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years(years)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{\itemize{
\item a vector of calendar years
}}
}
\value{
dataframe or NULL value
}
\description{
takes a vector of calendar years. For each year in the vector a dataframe is created for those that have data.
If data is not found for a particular year NULL is returned and a warning message is printed
}
\examples{
\dontrun{
  testyears <- c(2013,2014)
  fars_read_years(testyears)
}

}
