% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years(years)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{\itemize{
\item a vector of calendar years
}}
}
\value{
summaries of the yearly data for each month
}
\description{
takes a vector of calendar years. For each year that exists in the vector a summary of its
data is generated.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
}

}
