% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename(year)}
\usage{
make_filename(year)
}
\arguments{
\item{year}{\itemize{
\item calendar year to add to filename
}}
}
\value{
filename as character string
}
\description{
takes a year value and creates a filename by concatenating "accident_\%d.csv.bz2" where \%d is the year.
First function converts input year to an integer. Then outputs a character string as the newly created
filename.

Usage

make_filename(year)

Arguments
}
\examples{
   make_filename(2013)
   make_filename(2015)

}
